refactor: 모호한 infra 계층 패키지 이름을 infrastructure으로 변경
