+++ src/main/java/roomescape/infrastructure/payment/PaymentErrorResult.java
@@ -1,4 +1,4 @@
-package roomescape.infra.payment;
+package roomescape.infrastructure.payment;
 
 record PaymentErrorResult(String code, String message) {
 }
